refactor: 쿠폰 서비스 역할 분리
