+++ src/main/java/coupon/service/writer/CouponWriterService.java
@@ -1,10 +1,12 @@
-package coupon.service;
+package coupon.service.writer;
 
 import coupon.domain.Coupon;
 import coupon.repository.CouponRepository;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.cache.annotation.CachePut;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,14 +16,17 @@ import org.springframework.transaction.annotation.Transactional;
 public class CouponWriterService {
 
     private static final Logger log = LoggerFactory.getLogger(CouponWriterService.class);
+
     private final CouponRepository couponRepository;
 
     @Transactional
+    @CachePut(cacheNames = "coupon", key = "#coupon.id")
     public Coupon create(Coupon coupon) {
         return couponRepository.save(coupon);
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
+    @Cacheable(cacheNames = "coupon", key = "#id")
     public Coupon getCouponWithWriter(Long id) {
         log.info("Find coupon by id: {}", id);
